# internal/pkg/docs/swagger.json
{
  "components": {
    "schemas": {
...
      "dto.MediaUploadForm": {
        "properties": {
          "file": {
            "form": "file",
            "format": "binary",
            "type": "string"
          },
          "isPublic": {
            "example": false,
            "form": "isPublic",
            "type": "boolean"
          },
          "thumbnail": {
            "form": "thumbnail",
            "format": "binary",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "type": "object"
      },
      "dto.MediaUploadResponse": {
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "dto.ProfileImageUploadForm": {
        "properties": {
          "image": {
            "form": "image",
            "format": "binary",
            "type": "string"
          }
        },
        "required": [
          "image"
        ],
        "type": "object"
      },
      "dto.ProfileImageUploadResponse": {
        "properties": {
          "profile_image": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "dto.UserInfoResponse": {
        "properties": {
          "apple_id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "google_id": {
            "type": "string"
          },
          "has_password": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "is_email_verified": {
            "type": "boolean"
          },
          "is_privacy_policy_accepted": {
            "type": "boolean"
          },
          "is_profile_complete": {
            "type": "boolean"
          },
          "is_regulation_accepted": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "phone_number": {
            "type": "string"
          },
          "profession": {
            "type": "string"
          },
          "profile_image": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "errorhandler.HttpError": {
        "description": "This error is returned for specific failures within the API",
        "properties": {
          "caller": {
            "type": "string"
          },
          "code": {
            "type": "integer"
          },
          "message": {}
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "description": "Enter your API key in the format: x-api-key",
        "in": "header",
        "name": "x-api-key",
        "type": "apiKey"
      },
      "BearerAuth": {
        "description": "Enter your JWT token in the format: Bearer <token>",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "RESTful API for go-starter.",
    "title": "go-starter API",
    "version": "1.0"
  },
...
  "paths": {
...
    "/api/v1/media/upload": {
      "post": {
        "description": "Uploads a media file",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/dto.MediaUploadForm"
              }
            }
          },
          "description": "Multipart form data",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/dto.MediaUploadResponse"
                }
              }
            },
            "description": "Successfully uploaded media file"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "Invalid request format"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          }
        ],
        "summary": "Upload media",
        "tags": [
          "Media"
        ]
      }
    },
    "/api/v1/public/media-cleanup": {
      "delete": {
        "description": "Deletes a Unused media file",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "Invalid request or media ID"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "Media not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Delete media",
        "tags": [
          "Media"
        ]
      }
    },
...
    "/api/v1/user/me": {
      "get": {
        "description": "Get User Info",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/dto.UserInfoResponse"
                }
              }
            },
            "description": "User Info"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "Invalid request format"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get User Info",
        "tags": [
          "User"
        ]
      }
    },
...
    "/api/v1/user/upload-profile": {
      "post": {
        "description": "Upload a profile image (JPEG/PNG/WebP up to 5MB) to object storage.",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/dto.ProfileImageUploadForm"
              }
            }
          },
          "description": "Multipart form data",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/dto.ProfileImageUploadResponse"
                }
              }
            },
            "description": "Successfully uploaded profile image"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "Invalid request format or file"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorhandler.HttpError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          }
        ],
        "summary": "Upload profile image",
        "tags": [
          "User"
        ]
      }
    }
  },
  "openapi": "3.1.0"
}